progress commit - Selection interactor
