+++ lib/Selection/tests/Selection-test.js
@@ -1,34 +1,484 @@
 import React from 'react';
-import { describe, beforeEach, it } from '@bigtest/mocha';
+import { describe, afterEach, beforeEach, it } from '@bigtest/mocha';
 import { expect } from 'chai';
-import { Field } from 'redux-form';
+import sinon from 'sinon';
+import {
+  Button,
+  TextField,
+  including,
+  Keyboard,
+  HTML,
+  Selection as SelectionInteractor,
+  SelectionList as SelectListInteractor,
+  SelectionOption as SelectionOptionInteractor
+} from '@folio/stripes-testing';
 
 import { mountWithContext } from '../../../tests/helpers';
-import TestForm from '../../../tests/TestForm';
-
 import Selection from '../Selection';
-import SelectionInteractor from './interactor';
+import SingleSelectionHarness from './SingleSelectionHarness';
 
-describe('Selection', () => {
-  const selection = new SelectionInteractor();
+// import SelectInteractor from './interactor';
 
-  describe('coupled to redux-form', () => {
+describe.only('Selection, Single select', () => {
+  const selection = SelectionInteractor('test selection');
+
+  const expectClosedMenu = () => {
+    selection.has({ open: false });
+  };
+
+  const expectOpenedMenu = () => {
+    selection.has({ open: true });
+  };
+
+  const listOptions = [
+    { value: 'test0', label: 'Option 0' },
+    { value: 'test1', label: 'Option 1' },
+    { value: 'test2', label: 'Option 2' },
+    { value: 'sample0', label: 'Sample 0' },
+    { value: 'invalid', label: 'Sample 1' },
+    { value: 'sample2', label: 'Sample 2' }
+  ];
+
+  const testId = 'testingId';
+
+
+
+  beforeEach(async () => {
+    await mountWithContext(
+      <Selection
+        label="test selection"
+        id={testId}
+        dataOptions={listOptions}
+      />
+    );
+  });
+
+  it('renders the control', () => {
+    selection.exists();
+  });
+
+  it('does not have a value', () => {
+    selection.has({ value: '' });
+  });
+
+  it('renders the supplied id prop', () => {
+    selection.has({ id: testId });
+  });
+
+  it('list is hidden by default', expectClosedMenu);
+
+  it('control has the has-popup attribute "listbox"', () => {
+    Button({ hasPopup: 'listbox' });
+  });
+
+  it('control\'s aria-expanded attribute is false', expectClosedMenu);
+
+  it('control\'s aria-labelledBy attribute is set', () => {
+    Button({ ariaLabelledBy: `selected-${testId}-item` });
+  });
+
+  // stub test for legacy (tether prop) coverage - remove when tether prop is removed
+  describe('legacy rendering (tether prop)', () => {
+    const tetherSelect = SelectionInteractor('tether selection');
+    // the spy won't work if called within beforeEach
+    const spy = sinon.spy(console, 'error');
+    const text = 'Warning: The prop `tether` of `SingleSelect` is deprecated. Tether has been replaced by Popper. Use the /"popper/" prop to adjust the underlying Popper component.';
+
     beforeEach(async () => {
       await mountWithContext(
-        <TestForm>
-          <Field
-            name="testField"
-            component={Selection}
-            dataOptions={[
-              { value: 'test', label: 'Hello World' }
-            ]}
-          />
-        </TestForm>
+        <Selection
+          id={testId}
+          dataOptions={listOptions}
+          label="tether selection"
+          tether={{}}
+        />
       );
     });
 
+    afterEach(async () => {
+      spy.restore();
+    });
+
     it('renders the control', () => {
-      expect(selection.controlPresent).to.be.true;
+      tetherSelect.exists();
+    });
+
+    it('list is hidden by default', expectClosedMenu);
+
+    it('renders a deprecation warning', () => {
+      expect(spy.calledWith(text)).to.be.true;
+    });
+  });
+  // end stub test for legacy rendering
+
+  describe('supplying a label prop', () => {
+    beforeEach(async () => {
+      await mountWithContext(
+        <Selection
+          id={testId}
+          dataOptions={listOptions}
+          label="test selection"
+        />
+      );
+    });
+
+    it('renders the label', () => {
+      selection.exists();
+    });
+
+    it('control\'s aria-labelledBy attribute is set', () => {
+      Button({ ariaLabelledBy: `sl-label-${testId} selected-${testId}-item` });
+    });
+  });
+
+  describe('without options', () => {
+    beforeEach(async () => {
+      await mountWithContext(
+        <Selection
+          label="test selection"
+          id={testId}
+          dataOptions={[]}
+        />
+      );
+
+      await selection.toggle();
+    });
+
+    it('opens the list', expectOpenedMenu);
+  });
+
+  describe('supplying optionAlignment prop', () => {
+    describe('setting option alignment to "start"', () => {
+      beforeEach(async () => {
+        await mountWithContext(
+          <Selection
+            label="test selection"
+            dataOptions={listOptions}
+            optionAlignment="start"
+          />
+        );
+      });
+
+      it('should align options correctly', () => {
+        SelectionOptionInteractor().has({ className: including('optionStart') });
+      });
+    });
+
+    describe('setting option alignment to "end"', () => {
+      beforeEach(async () => {
+        await mountWithContext(
+          <Selection
+            label="test selection"
+            dataOptions={listOptions}
+            optionAlignment="end"
+          />
+        );
+      });
+
+      it('should align options correctly', () => {
+        SelectionOptionInteractor().has({ className: including('optionEnd') });
+      });
+    });
+
+    describe('setting option alignment to "center"', () => {
+      beforeEach(async () => {
+        await mountWithContext(
+          <Selection
+            label="test selection"
+            dataOptions={listOptions}
+            optionAlignment="center"
+          />
+        );
+      });
+
+      it('should align options correctly', () => {
+        SelectionOptionInteractor().has({ className: including('optionCentered') });
+      });
+    });
+
+    describe('setting option alignment to "outside"', () => {
+      beforeEach(async () => {
+        await mountWithContext(
+          <Selection
+            label="test selection"
+            dataOptions={listOptions}
+            optionAlignment="outside"
+          />
+        );
+      });
+
+      it('should align options correctly', () => {
+        SelectionOptionInteractor().has({ className: including('optionOutside') });
+      });
+    });
+  });
+
+  describe('clicking the control', () => {
+    beforeEach(async () => {
+      await selection.toggle();
+    });
+
+    it('opens the list', expectOpenedMenu);
+
+    it('focuses the filter input', () => {
+      TextField({ ariaLabel: including('filter'), focused: true }).exists();
+    });
+
+    it('filter input is rendered with "combobox" role', () => {
+      TextField({ ariaLabel: including('filter'), role: 'combobox' }).exists();
+    });
+
+    it(`list is rendered with ${listOptions.length} options`, () => {
+      SelectListInteractor({ optionCount: listOptions.length }).exists();
+    });
+
+    describe('clicking an option', () => {
+      beforeEach(async () => {
+        await selection.choose('Option 2');
+      });
+
+      it(`sets control value to ${listOptions[2].label}`, () => {
+        selection.has({ value: 'Option 2' });
+      });
+
+      it('closes the list', expectClosedMenu);
+
+      it('focuses the control/trigger', () => {
+        selection.has({ focused: true });
+      });
+    });
+
+    describe('filtering options', () => {
+      beforeEach(async () => {
+        await selection.filterOptions('Sample');
+      });
+
+      it('decreases list to 3 options', () => {
+        SelectListInteractor({ optionCount: 3 }).exists();
+      });
+
+      it('does not display the empty message', () => {
+        HTML({ role: 'alert' }).absent();
+      });
+
+      describe('reset filter', () => {
+        beforeEach(async () => {
+          await selection.filterOptions('');
+        });
+
+        it('should display initial list', () => {
+          SelectListInteractor({ optionCount: listOptions.length }).exists();
+        });
+      });
+
+      describe('clicking a filtered option', () => {
+        beforeEach(async () => {
+          await selection.choose('Sample 1');
+        });
+
+        it('sets the value appropriately', () => {
+          selection.has({ value: `${listOptions[5].label}` });
+        });
+      });
+
+      describe('No options available after filtering', () => {
+        beforeEach(async () => {
+          await selection.filterOptions('none');
+        });
+
+        it('displays the empty message', () => {
+          SelectionOptionInteractor('List is empty').exists();
+        });
+      });
+    });
+  });
+
+  describe('Selection, initial value', () => {
+    beforeEach(async () => {
+      await mountWithContext(
+        <SingleSelectionHarness
+          label="test selection"
+          initValue={listOptions[1].value}
+          options={listOptions}
+        />
+      );
+    });
+
+    it("renders the appropriate option's label", () => {
+      selection.has({ value: listOptions[1].label });
+    });
+
+    describe('Keyboard : letter press', () => {
+      beforeEach(async () => {
+        // press the S key...
+        await Keyboard.pressKey(83);
+      });
+
+      it('should set up the first filtered value', () => {
+        selection.has({ value: listOptions.find(o => o.value.startsWith('s')).label });
+      });
+    });
+
+    describe('Keyboard : down arrow on control', () => {
+      beforeEach(async () => {
+        await Keyboard.arrowDown();
+      });
+
+      it('opens the selection menu', expectOpenedMenu);
+
+      it('selects the next option', () => {
+        selection.has({ value: listOptions[2].label });
+      });
+    });
+
+    describe('Keyboard : up arrow on control', () => {
+      beforeEach(async () => {
+        // back twice to keep this from passing if the down arrow test fails.
+        await Keyboard.arrowUp();
+        await Keyboard.arrowUp();
+      });
+
+      it('opens the selection menu', expectOpenedMenu);
+
+      it('selects the previous option', () => {
+        selection.has({ value: listOptions[0].label });
+      });
+    });
+
+    describe('Keyboard: Enter press with open option list', () => {
+      beforeEach(async () => {
+        await selection.enter();
+      });
+
+      it('closes the list', expectClosedMenu);
+    });
+
+    describe('Keyboard: Enter press with closed option list', () => {
+      beforeEach(async () => {
+        await selection.enter();
+      });
+
+      it('opens the list', expectOpenedMenu);
+    });
+
+    describe('Filtering option list: cursor on first', () => {
+      beforeEach(async () => {
+        await selection.filterOptions('sam');
+      });
+
+      it.only('sets cursor to first result', () => {
+        SelectionOptionInteractor.has({ className: including('cursor'), index: 0 });
+      });
+
+      it('sets the appropriate aria-activedescendant on the filter', () => {
+        expect(selection.filterActiveDescendant).to.equal(selection.options(0).id);
+      });
+
+      describe('Keyboard : esc press', () => {
+        beforeEach(async () => {
+          await selection.pressKey('input', 27, true);
+        });
+
+        it('should close the list', expectClosedMenu);
+      });
+
+      describe('Keyboard : tab press', () => {
+        beforeEach(async () => {
+          await selection.pressKey('input', 9, true);
+        });
+
+        it('should close the list', expectClosedMenu);
+      });
+
+      describe('Keyboard : space press when filter input is empty', () => {
+        beforeEach(async () => {
+          await selection.fillFilter('');
+          await selection.pressKey('input', 32, true);
+        });
+
+        it('should close the list', expectClosedMenu);
+      });
+
+      describe('Keyboard control on filtered list: move cursor down', () => {
+        beforeEach(async () => {
+          await selection.moveToNextOption('input', true);
+          await selection.moveToNextOption('input');
+        });
+
+        it('sets cursor to third result', () => {
+          // expect(selection.options(0).isCursored).to.be.false;
+          expect(selection.options(2).isCursored).to.be.true;
+        });
+
+        it('sets the appropriate aria-activedescendant on the filter', () => {
+          expect(selection.filterActiveDescendant).to.equal(selection.options(2).id);
+        });
+
+        describe('Keyboard control on filtered list: move cursor up', () => {
+          beforeEach(async () => {
+            await selection.moveToPreviousOption('input');
+          });
+
+          it('sets cursor to second result', () => {
+            // expect(selection.options(2).isCursored).to.be.false;
+            expect(selection.options(1).isCursored).to.be.true;
+          });
+
+          it('sets the appropriate aria-activedescendant on the filter', () => {
+            expect(selection.filterActiveDescendant).to.equal(selection.options(1).id);
+          });
+
+          describe('Keyboard control on filtered list: pressing "Enter"', () => {
+            beforeEach(async () => {
+              await selection.pressEnter('input');
+            });
+
+            it('closes the option list', expectClosedMenu);
+
+            it('sets the cursored option as the value', () => {
+              expect(selection.valLabel).to.equal(listOptions[4].label);
+            });
+          });
+        });
+      });
+
+      describe('Supplied an \'error\' prop', () => {
+        beforeEach(async () => {
+          await mountWithContext(
+            <Selection
+              dataOptions={listOptions}
+              error="Selection is invalid!"
+            />
+          );
+        });
+
+        it('renders a validation message', () => {
+          expect(selection.inputError).to.be.true;
+        });
+
+        it('applies error style to control', () => {
+          expect(selection.hasErrorStyle);
+        });
+      });
+
+      describe('Supplied an \'warning\' prop', () => {
+        beforeEach(async () => {
+          await mountWithContext(
+            <Selection
+              dataOptions={listOptions}
+              warning="You might want to choose something different!"
+            />
+          );
+        });
+
+        it('renders a warning validation message', () => {
+          expect(selection.inputWarning).to.be.true;
+        });
+
+        it('applies \'warning\' style to control', () => {
+          expect(selection.hasWarningStyle);
+        });
+      });
     });
   });
 });
