+++ lib/Selection/tests/SingleSelectionHarness.js
@@ -5,6 +5,7 @@ import Selection from '../Selection';
 class SingleSelectionHarness extends React.Component {
   static propTypes = {
     initValue: PropTypes.string,
+    label: PropTypes.node,
     options: PropTypes.arrayOf(PropTypes.object),
   }
 
@@ -19,6 +20,7 @@ class SingleSelectionHarness extends React.Component {
   render() {
     return (
       <Selection
+        label={this.props.label}
         value={this.state.fieldVal}
         dataOptions={this.props.options}
         onChange={(val) => { this.setState({ fieldVal: val }); }}
